requirements.txt
Django==2.2
django-classy-tags==2.0.0
django-cms==3.9.0
django-formtools==2.3
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-snippet==3.0.0
djangorestframework==3.9.2
pytz==2021.1
sqlparse==0.4.1
